+++ test/common/test_sparse_reps.f
@@ -20,7 +20,7 @@
       subroutine test_csc_to_rsc()
       
       implicit real *8 (a-h,o-z)
-      integer, allocatable :: row_ptr(:),col_ind(:),col_ptr(:),
+      integer(8), allocatable :: row_ptr(:),col_ind(:),col_ptr(:),
      1   row_ind(:),iper(:)
 
       call prini(6,13)
